[create] Add menu router
