+++ src/components/Menu/Menu.css
@@ -1,0 +1,3 @@
+#link{
+    text-decoration: none;
+}
